+++ uguid/CHANGELOG.md
@@ -1,3 +1,5 @@
-# 0.5.0
+# 0.6.0
 
-Initial release.
+* Add `Guid::to_ascii_hex_lower` method. This is a const function that
+  creates a `[u8; 36]` array containing the GUID in standard
+  `xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx` format.